added spotify.py and scratch.py
