+++ younify/scratch.py
@@ -1,0 +1,27 @@
+#Bits of code no longer necessary
+
+class mywindow(QtWidgets.QMainWindow):
+    def __init__(self):
+        super(mywindow, self).__init__()
+        self.ui = Ui_MainWindow()
+        self.ui.setupUi(self)
+
+def RunMain_OBS():
+    app = QtWidgets.QApplication([])
+    file = QFile("resources\\dark.qss")
+    file.open(QFile.ReadOnly | QFile.Text)
+    stream = QTextStream(file)
+    app.setStyleSheet(stream.readAll())
+    application = mywindow()
+    application.show()
+    sys.exit(app.exec())
+
+class MyLogger(object):
+    def debug(self, msg):
+        pass
+
+    def warning(self, msg):
+        pass
+
+    def error(self, msg):
+        print(msg)
